refact: move keymap saving/loading away from keyedit.h/cpp
The save/load I/O for key mappings didn't really feel like it fit into the keyedit.h/cpp. Reason for this, I believe, is that the purpose of this file is more for defining the required UI elements for the keymap editor. Thus, including I/O in there introduces bit of a stretch to its responsibilities, violating the Single Responsibility Principle. However, going as far as introducing separate I/O adapter class seemed a bit overkill, so resorted to just moving the methods over to the mapping.h/cpp. As the KeyMappings class is now responsible for other keymap management anyways, it mostly makes sense to me to have I/O there, too.
